Renamed build -> build-commands and publish -> publish-commands to avoid conflict with .gitignore
