+++ scripts/release/build-commands/check-environment-variables.js
