+++ scripts/release/build-commands/check-npm-permissions.js
