+++ scripts/release/build-commands/run-automated-tests.js
